1.0 	修改：     Makefile 	修改：     README.md 	修改：     VERSION 	修改：     c 	修改：     c.c 	新文件：   c_help.c 	新文件：   conf/config.ini 	新文件：   config.c 	修改：     log/command.log 	修改：     pid/cprocess.pid 	新文件：   pid/s.pid 	删除：     pid/socket.pid 	修改：     s 	修改：     s.c 	修改：     s_daemon.c 	新文件：   s_help.c

+++ s.c
@@ -8,18 +8,19 @@
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <signal.h>
-//#include "popen.c"                                                            // 包含_popen()函数
+#include "config.c"
 
-#define	IP "0.0.0.0"                                                            // IP
-#define PORT 9606                                                               // 端口
-#define COMMAND_OUT_SIZE 1024
+
+#define COMMAND_OUT_SIZE 1024000
 #define PID_FILE_FLOW_SIZE 77
 #define BUFFER_SIZE 1024
-#define PID "pid/socket.pid"                                                    // 主进程PID存储文件
-#define COMMAND_OUT_FILE "log/command.log"                                               // 命令输出文件
 // 限制客服端IP
-#define I "127.0.0.1"
-#define P "192.168.99.109"
+#define LO "127.0.0.1"
+
+//#define IP "0.0.0.0"
+//#define PORT 9606
+//#define PID_FILE "pid/s.pid"
+//#define COMMAND_OUT_FILE "log/command.log"
 
 
 extern FILE *
@@ -27,48 +28,17 @@ _popen(const char *cmdstring, const char *type);
 
 extern void init_daemon(void);
 
-char help(void)
-{
-    static const char my_name[] = "socket";
-    static const char version[] = "version 1.0";
-    static const char subject[] = "socket";
-
-    static const struct {
-	const char *f, *s, *e;
-    } author = {
-    "aixiao@aixiao.me", "aixiao", "Author:",};
-    static const char usage[] = "Usage: socket [ip add] & [\"-f|-k file|pid\" || \"command\" ] | [-?h]";
-
-    static const char *s_help[] = {
-	"",
-	"    -?,-h     : this help",
-	"    -d        : daemon",
-	"    -f file   : Transfer file",
-	"    -k pid    : Child process PID",
-	"    -p,       : pid file(default: socket.pid)",
-	"",
-	"",
-	0
-    };
-
-
-    fprintf(stderr, "%s %s, %s\n", my_name, version, subject);
-    fprintf(stderr, "%s %s %s\n", author.e, author.s, author.f);
-    fprintf(stderr, "%s %s\n", my_name, usage);
-
-    int l;
-    for (l = 0; s_help[l]; l++)
-	fprintf(stderr, "%s\n", s_help[l]);
-}
+extern char help(void);
 
 int command(char buffer[])
 {
+	signed char *COMMAND_OUT_FILE = _config(4);
     FILE *fp = fopen(COMMAND_OUT_FILE, "a+");
 
     FILE *optput_buffer;                                                        // FILE文件操作的指针
-    char optput_flow[COMMAND_OUT_SIZE + 1];                                                    // 命令输出缓存
+    char optput_flow[COMMAND_OUT_SIZE + 1];                                     // 命令输出缓存
     optput_buffer = _popen(buffer, "r");
-    fread(optput_flow, sizeof(char), COMMAND_OUT_SIZE, optput_buffer);                         // 命令输出写入文件指针
+    fread(optput_flow, sizeof(char), COMMAND_OUT_SIZE, optput_buffer);          // 命令输出写入文件指针
     fseek(fp, 0, SEEK_END);
     fwrite(optput_flow, strlen(optput_flow), 1, fp);
 
@@ -79,8 +49,8 @@ int command(char buffer[])
 }
 
 void _kill(char buffer[]) {
-	printf("%s\n", buffer);
-	kill(atol(buffer), SIGKILL);
+    printf("%s\n", buffer);
+    kill(atol(buffer), SIGKILL);
 }
 
 int file(char buffer[], int client_socket, char *s_name)
@@ -89,9 +59,9 @@ int file(char buffer[], int client_socket, char *s_name)
 
 
     // 打开文件，准备写入
-    char *name = strdup(s_name);                                                      // 固定的文件名
-	printf("%s\n", name);
-	FILE *fp = fopen(name, "w");
+    char *name = strdup(s_name);                                                // 固定的文件名
+    printf("%s\n", name);
+    FILE *fp = fopen(name, "w");
     if (NULL == fp) {
         printf("File Can Not Open To Write\n");
         exit(1);
@@ -103,59 +73,78 @@ int file(char buffer[], int client_socket, char *s_name)
     while ((length =
         (int) recv(client_socket, buffer, BUFFER_SIZE, 0)) > 0) {
         if (fwrite(buffer, sizeof(char), length, fp) < length) {
-	        printf("File: %s Write Failed\n", name);
-	        break;
-	    }
+            printf("File: %s Write Failed\n", name);
+            break;
+    }
         allCount += length;
         bzero(buffer, BUFFER_SIZE);
     }
-	bzero(buffer, BUFFER_SIZE);
+    bzero(buffer, BUFFER_SIZE);
     // 接收成功后，关闭文件
     printf("Receive File Total: %dK\n",
         allCount);
     fclose(fp);
 }
 
-int pid(char *optarg)
+int pid(int s, char *optarg)
 {
-    char *start = "start";
+	signed char *PID_FILE = _config(3);
     char *stop = "stop";
-    char *reload = "reload";
     char *status = "status";
     char pid[PID_FILE_FLOW_SIZE];
-    char dev[PID_FILE_FLOW_SIZE];
+    char proc[PID_FILE_FLOW_SIZE];
 
     if(optarg == NULL) {
-        char *pid_file = PID;                                                   // 固定的文件名
-        FILE *fp = fopen(pid_file, "w");
-        fprintf(fp, "%d\n", getpid());
+        FILE *fp = fopen(PID_FILE, "w");
+        fprintf(fp, "%d", getpid());
         fclose(fp);
-    } else if(strcmp(optarg, stop) == 0) {
-        char *pid_file = PID;
-        FILE *fp = fopen(pid_file, "r");
-        fgets(pid, sizeof(pid), fp);
-        sprintf(dev, "/dev/%s", pid);
-        if(access(dev, 0) == 0)
-            printf("%s", pid);
-        int s_pid = atoi(pid);
-        kill(s_pid, SIGKILL);
+    }
+
+    if(s == 's'){
+        if(strcmp(optarg, stop) == 0) {
+            FILE *fp = fopen(PID_FILE, "r");
+            fgets(pid, sizeof(pid), fp);
+            sprintf(proc, "/proc/%s", pid);
+            if(access(proc, 0) == 0)
+                printf("%s\n", pid);
+            int s_pid = atoi(pid);
+            kill(s_pid, SIGKILL);
+            fclose(fp);
+        } else if(strcmp(optarg, status) == 0) {
+            FILE *fp = fopen(PID_FILE, "r");
+            fgets(pid, sizeof(pid), fp);
+            sprintf(proc, "/proc/%s", pid);
+            if(access(proc, 0) == 0)
+                printf("%s\n", pid);
+            fclose(fp);
+        }
+    } else if(s == 'p') {
+        FILE *fp = fopen(optarg, "w");
+        fprintf(fp, "%d\n", getpid());
         fclose(fp);
     }
 }
 
 int _main(int argc, char *argv[])
 {
+    signed char *IP = _config(0);
+    int PORT = atoi(_config(1));
+    signed char *CLIENTIP = _config(2);
+
     int ch;
     opterr = 0;
-    while((ch = getopt(argc, argv, "ds:h?")) != -1)
+    while((ch = getopt(argc, argv, "ds:p:h?")) != -1)
     switch (ch) {
     case 'd':
         init_daemon();
-        pid(0);
+        pid(0, 0);
         break;
     case 's':
-        pid(optarg);
+        pid('s', optarg);
         exit(0);
+        break;
+    case 'p':
+        pid('p', optarg);
         break;
     case 'h': case '?':
         help();
@@ -164,11 +153,6 @@ int _main(int argc, char *argv[])
     default:
         ;
     }
-
-	static const struct {
-		const char *i, *p;
-    } ip = {
-		I, P,};
 
     int server_socket = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);              // 创建套接字
     struct sockaddr_in serv_addr;                                               // 将套接字和IP、端口绑定
@@ -188,44 +172,44 @@ int _main(int argc, char *argv[])
 
         int len = sizeof(clnt_addr);
         getpeername(server_socket, (struct sockaddr *) &clnt_addr, &len);
-		char *clnt_ip = inet_ntoa(clnt_addr.sin_addr);
+        char *clnt_ip = inet_ntoa(clnt_addr.sin_addr);
         printf("other side IP：%s\n", inet_ntoa(clnt_addr.sin_addr));
         printf("other side PORT：%d\n", ntohs(clnt_addr.sin_port));
 
         memset(buffer, 0, sizeof(buffer));
         recv(client_socket, buffer, BUFFER_SIZE, 0);                            // 接收客户端发来消息
 		
-		if(strcmp(clnt_ip, ip.i) == 0 || strcmp(clnt_ip, ip.p) == 0) {
+        if(strcmp(clnt_ip, LO) == 0 || strcmp(clnt_ip, CLIENTIP) == 0) {
         if (buffer[0] == '-' && buffer[1] == 'f') {
-			int l = strlen(buffer);
-			for (int i = 0; i < l || buffer[i] == '\0'; i++) {
-				buffer[i] = buffer[i + 3]; 
-			}
-			char *name = strdup(buffer);
-			memset(buffer, 0, sizeof(buffer));
+            int l = strlen(buffer);
+            int i;
+            for (i = 0; i < l || buffer[i] == '\0'; i++) {
+                buffer[i] = buffer[i + 3]; 
+            }
+            char *name = strdup(buffer);
+            memset(buffer, 0, sizeof(buffer));
             file(buffer, client_socket, name);
-		}
-		else if (buffer[0] == '-' && buffer[1] == 'k') {
-			int l = strlen(buffer);
-			for (int i = 0; i < l || buffer[i] == '\0'; i++) {
-				buffer[i] = buffer[i + 3]; 
-			}
-			_kill(buffer);
+        } else if (buffer[0] == '-' && buffer[1] == 'k') {
+            int l = strlen(buffer);
+            int i;
+            for (i = 0; i < l || buffer[i] == '\0'; i++) {
+                buffer[i] = buffer[i + 3]; 
+            }
+            _kill(buffer);
         } else {
-			pid_t pid;
-			int status,i;
-			if(fork()==0){		// 子进程
-				command(buffer);
-				exit(1);
-			}
+            pid_t pid;
+            int status,i;
+            if(fork()==0){		// 子进程
+                command(buffer);
+                exit(1);
+            }
         }
-		}
-		else {
-			printf("%s\n", clnt_ip);
-		}
+        }
+        else {
+            printf("%s\n", clnt_ip);
+        }
         close(client_socket);                                                   // 关闭套接字
     }
-
     close(server_socket);
     return 0;
 }
